# docs/02-design-specs/api-contracts/openapi.yaml
openapi: 3.1.1
info:
  title: Vocabest Core API
  version: "1.0.0"
servers:
  - url: /api/v1
paths:
  /users:
    post:
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      operationId: listUsers
      deprecated: true
      description: "Deprecated: Collection GET is banned. Use GraphQL gateway instead."
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
  /users:whoami:
    get:
      operationId: whoami
      description: "Returns the authenticated user details based on GCP IAP header"
      parameters:
        - name: x-goog-authenticated-user-email
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /users:onboard:
    post:
      operationId: onboardUser
      description: "Registers and onboards a new user if not exists"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserOnboardRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /users/{userId}:
    get:
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: patchUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPatchRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{userId}:nextQuestion:
    post:
      operationId: getNextQuestion
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizQuestionResponse"
        "204":
          description: No Content (No more questions available)
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{userId}:nextErrorQuestion:
    post:
      operationId: getNextErrorQuestion
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizQuestionResponse"
        "204":
          description: No Content (No more error questions available)
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{userId}:errorReviewCount:
    get:
      operationId: getErrorReviewCount
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorReviewCountResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{userId}:submitAnswer:
    post:
      operationId: submitAnswer
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAnswerRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitAnswerResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{userId}:wordleTarget:
    post:
      operationId: getWordleTarget
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WordBankResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{userId}:validateWordleGuess:
    post:
      operationId: validateWordleGuess
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WordleValidationRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WordleValidationResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  schemas:
...
    SubmitAnswerRequest:
      type: object
      properties:
        questionId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174003"
        selectedDistractor:
          type: string
          example: "banana"
    SubmitAnswerResponse:
      type: object
      properties:
        isCorrect:
          type: boolean
          example: false
        correctAnswer:
          type: string
          example: "apple"
        explanationRootAffix:
          type: string
          example: "No specific root"
        explanationMnemonic:
          type: string
          example: "A for Apple"
    ErrorReviewCountResponse:
      type: object
      properties:
        count:
          type: integer
          example: 5
    UserOnboardRequest:
      type: object
      properties:
        email:
          type: string
          example: "user@example.com"
        targetLevel:
          type: string
          example: "JUNIOR_HIGH"
        dailyTargetQuestions:
          type: integer
          example: 20
    UserRequest:
      type: object
      properties:
        email:
          type: string
          example: "user@example.com"
        role:
          type: string
          example: "LEARNER"
        targetLevel:
          type: string
          example: "JUNIOR_HIGH"
        dailyTargetQuestions:
          type: integer
          example: 20
        learningStreak:
          type: integer
          example: 5
    UserPatchRequest:
      type: object
      properties:
        targetLevel:
          type: string
          example: "SENIOR_HIGH"
        dailyTargetQuestions:
          type: integer
          example: 30
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        email:
          type: string
          example: "user@example.com"
        role:
          type: string
          example: "LEARNER"
        targetLevel:
          type: string
          example: "JUNIOR_HIGH"
        dailyTargetQuestions:
          type: integer
          example: 20
        learningStreak:
          type: integer
          example: 5
        maxLearningStreak:
          type: integer
          example: 5
        maxDailyQuestions:
          type: integer
          example: 20
...
    WordleValidationRequest:
      type: object
      properties:
        guess:
          type: string
          example: "apple"
      required:
        - guess
    WordleValidationResponse:
      type: object
      properties:
        valid:
          type: boolean
          example: true
      required:
        - valid
...
    WordBankResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174002"
        word:
          type: string
          example: "apple"
        partsOfSpeech:
          type: string
          example: "n."
        chineseTranslation:
          type: string
          example: "蘋果"
        targetLevel:
          type: string
          example: "JUNIOR_HIGH"
        difficultyLevel:
          type: integer
          example: 1
        examFrequency:
          type: integer
          example: 5
...
    QuizQuestionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174003"
        wordBankId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174002"
        contextualCloze:
          type: string
          example: "I eat an ___ every day."
        chineseTranslation:
          type: string
          example: "我每天吃一顆蘋果。"
        correctAnswer:
          type: string
          example: "apple"
        distractor1:
          type: string
          example: "banana"
        distractor2:
          type: string
          example: "orange"
        distractor3:
          type: string
          example: "grape"
        explanationRootAffix:
          type: string
          example: "No specific root"
        explanationMnemonic:
          type: string
          example: "A for Apple"
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Invalid parameter format"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Resource not found"